% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{get.folds}
\alias{get.folds}
\title{get.folds}
\usage{
get.folds(n, nfolds)
}
\arguments{
\item{n}{Number of subjects}

\item{nfolds}{Number of folds for cross validation}
}
\description{
Returns folds of data for cross validation
}
