% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{group.sums}
\alias{group.sums}
\title{group.sums}
\usage{
group.sums(xx, groups)
}
\arguments{
\item{xx}{Vector of values to sum within groups}

\item{groups}{Vector indicating group membership}
}
\description{
Adds values within groups
}
