% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclCV.R
\name{mcl.cv}
\alias{mcl.cv}
\title{mcl.cv}
\usage{
mcl.cv(Z, W, y, mu = 1, groups, lam.seq = NULL, lam.max = NULL,
  nlam = c(10, 10), min.frac = 0.01, nfolds = 5, maxit = 10000,
  thresh = 1e-04, std = TRUE, verbose = 1)
}
\arguments{
\item{Z}{Otu matrix at group level}

\item{W}{Otu matrix within groups}

\item{y}{Outcome vector}

\item{mu}{Lagrange scaling parameter, default 1}

\item{groups}{Vector indicating group membership for each feature}

\item{lam.seq}{Sequence of lambda values, default NULL}

\item{lam.max}{Maximum lambda value, default NULL}

\item{nlam}{2-vector of number of lambdas to test (lam1 and lam2), default (10, 10)}

\item{min.frac}{Fraction of max.lam that the minimum lambda value should be, default 0.01}

\item{nfolds}{Number of folds for cross validation, default 5}

\item{maxit}{Maximum number of iterations, default 10000}

\item{thresh}{Threshold for convergence, default 0.0001}

\item{std}{Logical indicating whether data should be standardized, default TRUE}

\item{verbose}{If 0, prints nothing; if 1, prints which fold you're on; if 2, also prints a progress bar for lambdas. Default 1}
}
\description{
Cross-validated choice of lam1 and lam2 (over a grid of lambdas)
}
