% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl.R
\name{mcl}
\alias{mcl}
\title{mcl}
\usage{
mcl(Z, W, y, mu = 1, groups, lam.seq = NULL, lam.max = NULL,
  nlam = c(10, 10), min.frac = 0.1, maxit = 10000, thresh = 1e-04,
  std = TRUE, verbose = FALSE)
}
\arguments{
\item{Z}{Otu matrix at group level}

\item{W}{Otu matrix within groups}

\item{y}{Outcome vector}

\item{mu}{Lagrange scaling parameter; default 1}

\item{groups}{Vector of group membership per feature}

\item{lam.seq}{Sequence of lambda values; default NULL}

\item{lam.max}{Maximum lambda value; default NULL}

\item{nlam}{Vector of length 2 indicating number of lambda values to test for lam1 and lam2; default (10, 10)}

\item{min.frac}{Fraction of max.lam that the minimum lambda value should be; default 0.01}

\item{maxit}{Maximum number of iterations; default 10000}

\item{thresh}{Threshold for convergence; default 0.0001}

\item{std}{Logical indicating whether data should be standardized; default TRUE}

\item{verbose}{Print lambdas as indicator of progress?; default FALSE}
}
\description{
Calculates MCL coefficient estimates over a grid of (lam1, lam2) pairs
}
