% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mclC_step}
\alias{mclC_step}
\title{mclC}
\usage{
mclC_step(Z, W, y, facZ, facW, step, groups, groupIdx, mu, lam1, lam2, thresh,
  maxit)
}
\arguments{
\item{Z}{OTU matrix (group level)}

\item{W}{OTU matrix (within-group level)}

\item{y}{Vector of outcomes}

\item{facZ}{Scaling factors for Z}

\item{facW}{Scaling factors for W}

\item{step}{Step-size}

\item{groups}{Vector of group membership}

\item{groupIdx}{List of feature indices in each group}

\item{mu}{Lagrange scaling parameter}

\item{lam1}{Lambda for group-level lasso penalty}

\item{lam2}{Lambda for within-group lasso penalty}

\item{thresh}{Parameter defining convergence}

\item{maxit}{Maximum number of iterations allowed}
}
\value{
List of beta matrix, gamma matrix
}
\description{
Runs my multilevel compositional lasso
}
