% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{softC2}
\alias{softC2}
\title{softC2}
\usage{
softC2(x, lam)
}
\arguments{
\item{x}{Vector whose elements should be soft thresholded}

\item{lam}{Parameter by which to soft threshold}
}
\description{
Soft thresholds each element of a vector x by lambda
}
