% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcl_dataprep.R
\name{dataprep.mcl}
\alias{dataprep.mcl}
\title{dataprep.mcl}
\usage{
dataprep.mcl(otu.table, groups, rare.count = 1, rare.prop = 0,
  pseudocount = 0.5)
}
\arguments{
\item{otu.table}{Matrix of otu counts; rows are subjects, columns are OTUs.}

\item{groups}{Vector indicating group membership}

\item{rare.count}{OTU count considered "rare"; OTUs with total read count <= rare.count will be excluded. Default is 1 (exclude singletons). If <= 0, no filter is applied for rare OTUs based on counts.}

\item{rare.prop}{Alternative to rare.count; minimum proportion of samples in which the OTU should appear. Default is 0. If <= 0, no filter is applied for rare OTUs based on proportions.}

\item{pseudocount}{Count to replace zeros; default is 0.5.}
}
\value{
Returns a list with elements Z, W, and groups for input into MCL function, and a vector with the names of groups with only one element (excluded from W matrix).
}
\description{
Prepares input matrices for MCL function based on otu count matrix.
Rows of the otu table should be subjects.
}
