% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/service.R
\name{dataprep.mcl}
\alias{dataprep.mcl}
\title{dataprep.mcl}
\usage{
dataprep.mcl(otu.table, groups)
}
\arguments{
\item{otu.table}{Matrix of otu counts; rows are subjects, columns are OTUs.}

\item{groups}{Vector indicating group membership}
}
\value{
Returns a list with elements Z and W for input into MCL function
}
\description{
Prepares input matrices for MCL function based on otu count matrix.
Rows of the otu table should be subjects.
}
